Optimized NumPy Code, and Pythran code, Removed unnecessary files.
